% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion_payer.R
\name{model_conversion_rpp}
\alias{model_conversion_rpp}
\title{Bayesian model that combines conversion estimation and value estimation for converted cases}
\usage{
model_conversion_rpp(
  alpha,
  beta,
  success,
  total_sample_size,
  shape,
  rate,
  sum_sample,
  n_post = 1e+05
)
}
\arguments{
\item{alpha}{Parameter for prior in bernoulli-beta model representing number of success - 1}

\item{beta}{Parameter for prior in bernoulli-beta model representing number of fails - 1}

\item{success}{Actual number of successful cases (payers) in your data.}

\item{total_sample_size}{Total number of cases in your data.}

\item{shape}{Parameter for prior in exponential-gamma model representing the number of samples from before - 1}

\item{rate}{Parameter for prior in exponential-gamma model representing the sum of samples from before}

\item{sum_sample}{Sum of variable (money) for each successful case.}

\item{n_post}{Size of sample from posterior distribution}
}
\value{
List of 3 posterior distributions: post_conv representing conversion rate (success rate) from bernoulli-beta model,
 post_rpp representing value of variable (money) per successful case from exponential-gamma model, post_rpu representing value
 of variable per each case (money per all cases), post_rpu = post_conv * post_rpp
}
\description{
Model uses total number of cases and successful cases to estimate conversion based on bernoulli-beta model
and for successful cases estimates continuous variable using exponential-gamma model
}
\details{
Final result is estimated value for each case (not just converted cases). Motivation is when you have some percent of
users that spend money and others that spend no money. First we estimate conversion to payers rate, and for payers estimate
revenue per payer (rpp). Final posterior distribution is multiple of these two posterior distributions and is representing
revenue per user (rpu). 

Check model_bernoulli_beta and model_exponential_gamma for details regarding each model.
}
\author{
Elio Bartoš
}
